Fixed performance of 1. removing large numbers of tracks from a mix 2. automixing large number of tracks in a mix. This was happening because it was recalculating recommendations for every single item every time you did something. Very slow.

+++ src/MixPlanner/Commands/RemoveTracksFromMixCommand.cs
@@ -8,11 +8,14 @@ namespace MixPlanner.Commands
     public class RemoveTracksFromMixCommand : CommandBase<IEnumerable<IMixItem>>
     {
         readonly IMix mix;
+        readonly IDispatcherMessenger messenger;
 
-        public RemoveTracksFromMixCommand(IMix mix)
+        public RemoveTracksFromMixCommand(IMix mix, IDispatcherMessenger messenger)
         {
             if (mix == null) throw new ArgumentNullException("mix");
+            if (messenger == null) throw new ArgumentNullException("messenger");
             this.mix = mix;
+            this.messenger = messenger;
         }
 
         protected override bool CanExecute(IEnumerable<IMixItem> parameter)
@@ -22,7 +25,21 @@ namespace MixPlanner.Commands
 
         protected override void Execute(IEnumerable<IMixItem> parameter)
         {
-            mix.RemoveRange(parameter);
+            using (mix.Lock())
+            {
+                // Optimization: if we are nuking the whole mix (all tracks
+                // selected), it is quicker to clear than remove individual tracks.
+                if (parameter.Count() == mix.Count)
+                {
+                    mix.Clear();
+                    return;
+                }
+
+                // Have to disable recommendations otherwise it recalcs
+                // recommendations every time you delete a row.
+                using (new DisableRecommendationsScope(messenger))
+                    mix.RemoveRange(parameter);
+            }
         }
     }
 }
